learnig testing the FastApi

+++ FastAPI_The_Complete_Course_2025/TodoApp/routers/todos.py
@@ -1,13 +1,18 @@
 from fastapi import  Depends, HTTPException, status, Path, APIRouter
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
-from models import Todos
-from database import get_db
+from ..models import Todos
+from ..database import get_db
 from typing import Annotated
+from .auth import get_current_user
 
-router = APIRouter(tags=["Todos"])
+router = APIRouter(
+    prefix="/todo",
+    tags=["Todos"]
+)
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title : str = Field(min_length=3)
@@ -15,32 +20,40 @@ class TodoRequest(BaseModel):
     priority : int = Field(gt=0, lt=6)
     complete : bool
     
+@router.get("/", status_code=status.HTTP_200_OK)
+def read_all_todo(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
-@router.get("/todo", status_code=status.HTTP_200_OK)
-def read_all_todo(db: db_dependency):
-    return db.query(Todos).all()
-
-@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-def read_todo_by_id(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.get("/{todo_id}", status_code=status.HTTP_200_OK)
+def read_todo_by_id(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail="Todo not found")
+    raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail="Todo not found.")
 
-@router.post("/todo/create_todo", status_code = status.HTTP_201_CREATED)
-def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+@router.post("/create_todo", status_code = status.HTTP_201_CREATED)
+def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
     
-@router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-def update_todo(db: db_dependency,
+@router.put("/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
+def update_todo(user: user_dependency, 
+                db: db_dependency,
                 todo_request: TodoRequest, 
                 todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id)
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id"))
     todo_post = todo_model.first()
     if todo_post is None:
-        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail = "Todo not found")
+        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail = "Todo not found.")
     # todo_model.title = todo_request.title
     # todo_model.description = todo_request.description
     # todo_model.priority = todo_request.priority
@@ -49,10 +62,12 @@ def update_todo(db: db_dependency,
     # db.add(todo_model)
     db.commit()
     
-@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.delete("/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found.")
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
